PQSC driver added
#### **The following changes are made in `.ini` file:**
##### **1) New parameters and options are added to `Setup` section:**
* Revisions - Data Server Version
* Revisions - Firmware Version
* Revisions - FPGA Version
* Preset - Factory Reset
* PQSC - Connect to PQSC
##### **2) Following options are hidden when connected to PQSC:**
* Device - Reference Clock
* Digital I/O - Mode
* Digital I/O - Output
* Digital I/O - Drive
* Digital I/O - Ext Clock
##### **3) `QA Monitor - Inputs` are updated:**
`QA Monitor - Input 1` and `QA Monitor - Input 2` are shown
automatically in measurement window.
##### **4) `Trigger Mode` of `Seqeuncer` section is updated:**
* `External Trigger` is replaced with `Receive Trigger`
* `ZSync Trigger` trigger mode is added
##### **5) `Rerun` option is added to `Sequencer` section**
##### **6) `Digital I/O - Mode` updated:**
* `Normal` options is renamed to `Manual` as in LabOne GUI.
* `QA Result QCCS` option is added.
##### **7) `Digital I/O - Ext Clock` updated:**
* `Internal 50 MHz` option is added.
##### **8) Following options are added to `Sequencer` section:**
* Sequencer - Strobe Index
* Sequencer - Strobe Slope
* Sequencer - Valid Index
* Sequencer - Valid Polarity
Note that these options are automatically hidden if the selected trigger
mode is `ZSync Trigger`
#### **The following changes are made in `.py` file:**
##### **1) `performOpen` method updated:**
The revision parameters are read from the device and the Labber
controller is updated in the beginning, right after connection is
established.
##### **2) `factory_reset` and `connect_to_pqsc` methods are added:**
These methods call the necessary functions from zhinst-toolkit.
##### **3) `version_parser` method is added:**
This method is used to parse the data server version string.
##### **4) `performArm` method is updated:**
`External Trigger` is replaced with `Receive Trigger` and
`ZSync Trigger` is added to the arm condition.
##### **5) `update_labber_controller` method is added:**
This method is used to read a quantity from the device and update the
Labber controller with the current value obtained.
##### **6) `update_sequencers` method is updated:**
The sequencer parameters should be set when the trigger mode
selected is `ZSync Trigger` even when the sequence type is `None`.